requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements-dev.in'
#
alabaster==1.0.0
    # via sphinx
argcomplete==3.6.2
    # via fusesoc
asgiref==3.8.1
    # via django
astroid==3.3.11
    # via pylint
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
    #   simplesat
babel==2.17.0
    # via sphinx
build==1.2.2.post1
    # via pip-tools
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
chardet==5.2.0
    # via diff-cover
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   pylint
    #   pytest
    #   sphinx
coverage[toml]==7.10.6
    # via pytest-cov
cryptography==45.0.4
    # via pyjwt
deprecated==1.2.18
    # via pygithub
diff-cover==9.6.0
    # via -r requirements-dev.in
dill==0.4.0
    # via pylint
django==5.2.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-spectacular
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   drf-spectacular
docutils==0.21.2
    # via
    #   sphinx
    #   sphinx-rtd-theme
drf-spectacular==0.28.0
    # via -r requirements.in
edalize==0.6.1
    # via fusesoc
factory-boy==3.3.3
    # via -r requirements-dev.in
faker==37.6.0
    # via factory-boy
fastjsonschema==2.21.1
    # via fusesoc
fusesoc==2.4.4
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via -r requirements.in
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via pylint
jinja2==3.1.6
    # via
    #   diff-cover
    #   edalize
    #   pytest-html
    #   sphinx
jsonschema==4.24.0
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
okonomiyaki==3.0.0
    # via simplesat
packaging==25.0
    # via
    #   build
    #   pytest
    #   sphinx
pip-tools==7.5.0
    # via -r requirements-dev.in
platformdirs==4.3.8
    # via pylint
pluggy==1.6.0
    # via
    #   diff-cover
    #   pytest
    #   pytest-cov
pycparser==2.22
    # via cffi
pygithub==2.6.1
    # via -r requirements.in
pygments==2.19.2
    # via
    #   diff-cover
    #   pytest
    #   sphinx
pyjwt[crypto]==2.10.1
    # via pygithub
pylint==3.3.8
    # via
    #   -r requirements-dev.in
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.6.1
    # via -r requirements-dev.in
pylint-plugin-utils==0.9.0
    # via pylint-django
pynacl==1.5.0
    # via pygithub
pyparsing==3.2.3
    # via fusesoc
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.1
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-html
    #   pytest-metadata
    #   pytest-mock
pytest-cov==6.2.1
    # via -r requirements-dev.in
pytest-django==4.11.1
    # via -r requirements-dev.in
pytest-html==4.1.1
    # via -r requirements-dev.in
pytest-metadata==3.1.1
    # via pytest-html
pytest-mock==3.14.1
    # via -r requirements-dev.in
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   drf-spectacular
    #   fusesoc
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   pygithub
    #   sphinx
roman-numerals-py==3.1.0
    # via sphinx
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
semver==3.0.4
    # via -r requirements.in
simplesat==0.9.2
    # via fusesoc
six==1.17.0
    # via simplesat
smmap==5.0.2
    # via gitdb
snowballstemmer==3.0.1
    # via sphinx
sphinx==8.2.3
    # via
    #   -r requirements.in
    #   sphinx-autodoc-typehints
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-autodoc-typehints==3.2.0
    # via -r requirements.in
sphinx-rtd-theme==3.0.2
    # via -r requirements.in
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlparse==0.5.3
    # via django
tomlkit==0.13.3
    # via pylint
typing-extensions==4.14.0
    # via pygithub
tzdata==2025.2
    # via
    #   django
    #   faker
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.4.0
    # via
    #   pygithub
    #   requests
wheel==0.45.1
    # via pip-tools
wrapt==1.17.2
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
